// frontend/lib/api/generated/model/productoOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */
import type { UnidadMedidaOut } from './unidadMedidaOut';

export interface ProductoOut {
  producto_id: number;
  codigo: string;
  nombre: string;
  categoria: string | null;
  alimento_id: number | null;
  perecible: boolean;
  dias_vida_util: number | null;
  stock_minimo_referencial: number | null;
  estado: string;
  unidad: UnidadMedidaOut;
}